+++ src/outputs/outputs.cpp
@@ -117,19 +117,31 @@ Outputs::Outputs(Mesh *pm, ParameterInput *pin, SimTime *tm) {
       op.block_name.assign(pib->block_name);
 
       Real dt = 0.0; // default value == 0 means that initial data is written by default
-      // for temporal drivers, setting dt to tlim ensures a final output is also written
+      int dn = -1;
       if (tm != nullptr) {
-        dt = pin->GetOrAddReal(op.block_name, "dt", tm->tlim);
+        dn = pin->GetOrAddInteger(op.block_name, "dn", -1.0);
+
+        // If this is a dn controlled output (dn >= 0), soft disable dt based triggering
+        // (-> dt = -1), otherwise setting dt to tlim ensures a final output is also
+        // written for temporal drivers.
+        const auto tlim = dn >= 0 ? -1 : tm->tlim;
+        dt = pin->GetOrAddReal(op.block_name, "dt", tlim);
       }
       // if this output is "soft-disabled" (negative value) skip processing
-      if (dt < 0.0) {
+      if (dt < 0.0 && dn < 0) {
         pib = pib->pnext; // move to next input block name
         continue;
       }
+
+      PARTHENON_REQUIRE_THROWS(!(dt >= 0.0 && dn >= 0),
+                               "dt and dn are enabled for the same output block, which "
+                               "is not supported. Please set at most one value >= 0.");
       // set time of last output, time between outputs
       if (tm != nullptr) {
         op.next_time = pin->GetOrAddReal(op.block_name, "next_time", tm->time);
         op.dt = dt;
+        op.next_n = pin->GetOrAddInteger(op.block_name, "next_n", tm->ncycle);
+        op.dn = dn;
       }
 
       // set file number, basename, id, and format
@@ -259,8 +271,6 @@ Outputs::Outputs(Mesh *pm, ParameterInput *pin, SimTime *tm) {
       if (op.file_type == "hst") {
         pnew_type = new HistoryOutput(op);
         num_hst_outputs++;
-      } else if (op.file_type == "vtk") {
-        pnew_type = new VTKOutput(op);
       } else if (op.file_type == "ascent") {
         pnew_type = new AscentOutput(op);
       } else if (op.file_type == "histogram") {
@@ -438,9 +448,19 @@ void Outputs::MakeOutputs(Mesh *pm, ParameterInput *pin, SimTime *tm,
   OutputType *ptype = pfirst_type_;
   while (ptype != nullptr) {
     if ((tm == nullptr) ||
-        ((ptype->output_params.dt >= 0.0) &&
-         ((tm->ncycle == 0) || (tm->time >= ptype->output_params.next_time) ||
-          (tm->time >= tm->tlim) || (signal == SignalHandler::OutputSignal::now) ||
+        // output is not soft disabled and
+        (((ptype->output_params.dt >= 0.0) || (ptype->output_params.dn >= 0)) &&
+         // either dump initial data
+         ((tm->ncycle == 0) ||
+          //  or by triggering time or cycle based conditions
+          ((ptype->output_params.dt >= 0.0) &&
+           ((tm->time >= ptype->output_params.next_time) ||
+            (tm->tlim > 0.0 && tm->time >= tm->tlim))) ||
+          ((ptype->output_params.dn >= 0) &&
+           ((tm->ncycle >= ptype->output_params.next_n) ||
+            (tm->nlim > 0 && tm->ncycle >= tm->nlim))) ||
+          // or by manual triggers
+          (signal == SignalHandler::OutputSignal::now) ||
           (signal == SignalHandler::OutputSignal::final) ||
           (signal == SignalHandler::OutputSignal::analysis &&
            ptype->output_params.analysis_flag)))) {